test: opengl: add test for filters
This test is able to use any offscreen opengl implementation available
to run the opengl filters engine on both GL Core and GL ES, and will
skip the test for one of those targets if no implementations are
available.

It currently mainly smoke-test the filters engine, while checking that
orientation handling is correct, and uses glReadPixels to check the
resulting rendered picture in a multi-platform way.