misc: es_format: set orientation during setup
During video_format_Setup(), the size, visible size and source aspect
ratio is defined for the given video_format_t instance, but those values
only make sense for a given orientation. Transforming the picture would
mean transform the coordinates, as seen in video_format_TransformBy().

To ensure the correct behaviour is found when transforming after the
setup, initialize the orientation to ORIENT_NORMAL as well.