package/macos: add an option to select the (lib)VLC license
Some contribs may be disabled depending on the license chosen.

By default we build with GPLv3 compatibility. libvlc-only builds may want
to restrict to LGPL v3 with ad-clause or LGPL v2.1 with ad-clause.

faad is automatically picked when building for GPL and not picked
otherwise. We should never force it.