vout: stop even when display is missing
`vout_Stop` should be called even if the display is missing as it is
also charged to disable the window if its still open. The display NULL
check is done anyway at the correct place in `vout_Stop` itself.

This fixes an assert with the `display` stream output that never closes
the window because `input_resource_StopFreeVout` is not enforced
compared to the classic playback scenario.

Here's the assert fixed by this patch:

```
vlc: ../../src/video_output/video_output.c:1948: vout_Release: Assertion `!sys->window_enabled' failed.
[5] from 0x00007ffff6c5f29c in vout_Release+1019 at ../../src/video_output/video_output.c:1948
[6] from 0x00007ffff6c5ee9a in vout_Close+698 at ../../src/video_output/video_output.c:1926
[7] from 0x00007ffff6bce970 in DestroyVout+840 at ../../src/input/resource.c:159
[8] from 0x00007ffff6bd0609 in input_resource_Release+99 at ../../src/input/resource.c:334
[9] from 0x00007fffd89b6e68 in Close+274 at ../../modules/stream_out/display.c:265
```