text_renderer: sapi: fix assert in RenderText
vout_subpictures:RenderText() now expects the subtitle to have been
rasterized, but we'll never render any text, so we want to signal that
the region must be discarded.

This is a bit inadequate since it means that text OSD and subtitles are
not displayed at all but we probably want to refactor this part later by
adding a new capability for text parsing before rendering.