dxva: use the actual surface size to test decoder support
The container may lie but the size that libavcodec requests is the one it
will use. We need this size to probe the decoder in D3D11. There doesn't
seem to be a way to check the size support in dxva2.

Similar to e4cc2f846b020e4e5a89046669c1eaddbebe0a34 but without setting
an output video_format_t.