sout: enforce single frames in `pf_send`
The stream output API is unclear on chained frame support. Currently,
modules decide if they want to support them or not.
As it seems that the current consensus is to support single frames only,
this assert will help notifying and fixing inconsistents behavior on
that regard.

Note that it could be perfectly valid to pass frames one by one from a
chain without having to NULL every `p_next` entry for each elements so
this assert might be deleted once it is clear that all stream output
modules are fixed to support single frames only.