sout: duplicate: stop storing NULL dup_ids
Storing NULL duplicated IDs was necessary before we had a reference to
the duplicate stream output stored with the duplicated id. Storing NULL
allowed to have an index equivalence between the stream array and the ID
array.

Now that IDs can be independently stored, lets remove the NULL pushes
and the unneeded NULL checks!