CI: update debian image
The libc++(abi)-dev packages have been removed because one of their
dependencies conflicted with one of the gstreamer dependencies. As a
result, the gstreamer dependency was not installed and pkg-config was
unable to satisfy the requirements for gstreamer. With the conflicting
package now gone, gstreamer is installed properly and the Debian CI job
is now able to build the gstdecode module.

Amends d0c0a3147a1f3d217cc59bfbd896f6ff5724905d.