meson: add rust (-Drust=enabled) option
This adds a -Drust feature option to the meson build system, which aims
to enable pure Rust modules.

This option will check for the presence of cargo (Rust package manager
and build system) and give a warning about the experimental nature of
the Rust modules.

This option is disabled by default.