rust: add binary helper crate for generating raw bindings
This binary crate, called vlcrs-sys-generator, uses the bindgen[1] crate
to automatically generate raw (extern "C"...) bindings, which are then
used by safe abstractions.

The goal is for this program to be used to generate raw bindings that
are which will be used by safe abstractions, each safely wrapping a part
of the VLC core API, generally a header will be a crate. This is done so
that the so that the dependency chain of the Rust modules closely
follows that of the C modules, in particular a demux module should not
need to be rebuild if it doesn't depend on vout.

Since writing and maintaining raw bindings is tedious and error-prone,
the raw bindings (extern "C"...) are generated using a Clang front-end,
bindgen, written in Rust. Updating the bindings should be as simple as
running `cargo run -p vlcrs-sys-generator'.

[1]: https://rust-lang.github.io/rust-bindgen/