dbus: dbus: refactor release of events
Tracklist events were aggregated by the processing code into it's own
linked-list which was then used to drpo the events.

The code was different from how events are released when closing the
application with events that were not sent to the dbus thread. This
commit unify the way they are released by duplicating the existing
release code matching the signal type and dropping them from the
ProcessEvents function instead.