dbus: dbus: delete event on error
When a second tracklist (append or remove) event was queued to the dbus
thread, it detected that an existing event was already there and
discarded the event without destroying it, despite the ownership being
transferred.

By checking whether the event was transferred or not and release it if
not, we don't risk leaking the event structures and the underlying items
hold by them.

No functional changes since those events were not discarded anymore.