httpd: use size_t for buffer size results
The value can never be negative and it simplifies code in places.

We also need to change the i_body which is used as a pointer in some cases.

The value can never be negative and it simplifies code in places.

We also need to change the i_body which is used as a pointer in some cases.