codec: avcodec: fix mask mappings for 32 bits RGB
The macros can not handle alpha channel properly,
as it does not only requires reassigning channels
but shifting.

Mapping to/from RGB32 also loses alpha information.

We can not rely on AV native macros for this and
need to declare chroma with fixed alpha position

Could also modify the macro to add left or right
shifts parameters and even less readable for 2 chromas.