vout: create the private pool independently from the display pool
Since we reserved pictures, the bigger size of the pool is not a benefit.
It was necessary in 3.0 where the pool was coming from the display module.
We can split them since they always worked on different pictures.

vd->source is the same video format as osys->display_fmt used for the
display pool in this case since there is no converter.