wingdi: use less hacky way to set RGB values
BITMAPINFO is actually BITMAPINFOHEADER and a dummy RGBQUAD.
It's better to have our own with the 3 fields. The red/green/blue fields were
not used, just taking enough space in the structure.

We could pass the proper sizeof in the BITMAPINFOHEADER when the 3 values
are used. It works with it on my machine, but it may not on older Windows
or older GPU drivers.