video_output: rename do_dr_spu
This name is confusing with the video direct rendering. It is more useful
to know that the display module will do the blending or not.

There is no early or late SPU blending. The key factor is whether the
(local) blending is done before the converter or after the converter.