test/libvlc: meson.build: add missing tests with deps
Add the tests that were missing from Makefile.am to meson.build, now
that we can define their dependencies.

Tests like test_libvlc_core and test_libvlc_media are depending on the
full list of plugins so that they are run after the plugins they use are
actually built. This list could be reduced in the future but since they
depend on platform plugins or a huge list of different plugins, it's
easier to test them right now after everything is built.