test: meson.build: add module_depends to tests
Add a field in tests to ensure they are built after the plugins that
needs them. It references the name and not the target directly to ease
its definition, and can also serve different purpose like potentially
linking the plugins directly in the future depending on the build
environment.

When a dependency to an inexistant module is added, the following error
will be triggered:

    ../test/meson.build:39:12: ERROR: Problem encountered: Test entry
    'test_libvlc_media_player' cannot depend on inexistant VLC plugin
    'nonexistantmodule'