modules: meson.build: name plugins targets
Ensure plugins targets can be referenced by naming them through the
`vlc_plugins_targets` dictionary. They can be used directly through

    vlc_plugins_targets['module_name']

where `module_name` is the 'name' field added in the `vlc_modules` dict.

In particular, tests can reference those targets to ensure the plugin is
built when running it. To do so, the current only way to achieve this is
to declare a dependency on the library as if it was linked, and then
get a partial dependency to remove the library from the link. For
instance:

    declare_dependency(link_with: vlc_plugins_targets['demux_mock'])
        .partial_dependency()