opengl: sampler: add preambles to vlc_gl_sampler
This allows vlc_gl_sampler to be used with the appropriate version of
GLSL enabled, which allows the use of more up-to-date GLSL features.

To avoid unnecessary complicated preprocessor interventions, we limit
the upper version of desktop GLSL to 410, which is the last version that
retains compatibility for attribute/varying declarations.