src/Makefile: fix emscripten library object conflict
Several posix related headers are added when they are compiled
for EMSCRIPTEN. However these headers are already served from
other parts of code, mainly with !HAVE_DARWIN and !OS2. Which
implies If it is not compiled for darwin (which is true for
emscripten) some posix libraries are added. And when the
emscripten section of the code adds them again, it creates a clash.

Fix this problem by removing such double addition of sources from
emscripten.