Makefile.am: don't ignore user distcheck options
The Automake documentation [1] states that the user could provide
DISTCHECK_CONFIGURE_FLAGS, and that AM_DISTCHECK_CONFIGURE_FLAGS
should provide additional flags that might be needed. So change
DISTCHECK_CONFIGURE_FLAGS for AM_DISTCHECK_CONFIGURE_FLAGS, because
as it stands, the user-provided DISTCHECK_CONFIGURE_FLAGS are
ignored and overwritten by the Makefile.am.

1. https://www.gnu.org/software/automake/manual/html_node/Checking-the-Distribution.html