caopengllayer: restore vout_display_SendEventDisplaySize
VLC 4.0 doesn't need that but 3.0 still has no windowing module, leading
to display size not being reported to the subtitle text renderer, and
leading to blurry subtitles and OSD.

Because the size is coming from the display, it means that the first
time the subtitle is displayed, it will always be blurry depending on
the real display size and original media size.

Regression from 8ff56952174af8e7a7d4ae61a413f88317c948a1.

Fixes #27793