extras: ci: remove --disable from distcheck
Taking the same ideas as b46611fee3c8a0f18b12120dee3886ceaddae7b6, we
can ensure the distcheck builds everything when run in the CI, so that
the currently disabled modules are built for real and tests can use
them. They should stay disabled since they would break distcheck when
the mandatory plugins cannot be built, but we can ensure that in the CI
in a reliable way.

We use it as an argument to make so that make actually erase the value
in the makefile.

Refs #27761 but medialibrary still needs to be installed on the platform
since contribs are not enabled with the current distcheck configuration.