share: Makefile.am: fix distcheck for lua
Distcheck was never tested with lua enabled, and it doesn't clean the
compiled lua files correctly in distcleancheck. There was also some
redundancy in how the scripts were defined.

This commit refactor the declaration in three separate variables:
 - LUA_EXTENSIONS, which is where new lua scripts should probably go.
 - LUA_EXTENSIONS_DOC, which include files that are shipped into the doc
   folder but not compiled.
 - LUA_EXTENSIONS_PKGDATA, which include files that are shipped into the
   pkgdata folder, and mostly consists of http/custom.lua now for some
   reason.

The .luac files are not included in the dist archive, and thus
can be automatically clean them.