config: dirs: implement VLC_USERDATA_PATH override
If the variable VLC_USERDATA_PATH is set, completely override the output
of config_GetUserDir for VLC_USERDATA_DIR to not output the platform
path, but the specified path instead.

This will allow overriding this variable in a VLC-specific manner,
without relying on mechanism like XDG dir overrides that might not be
implemented for every platforms.

Refs #28343