test: lua: use VLC_USERDATA_PATH
Darwin platforms are not using XDG_DATA_DIR. With the override done in
src/config/dirs.c, every platform (now and then) will use the content of
VLC_USERDATA_PATH instead to create files.

It also changes the way the Lua extension dir is set to be simpler to
configure in buildsystem (TOP_SRCDIR is already forwarded) and resilient
to working directory change.

Refs #28343