text_renderer: sapi: use new/delete for sys
We use a std::unique_ptr<> to handle the lifecycle of the sys pointer
with the error cases at the same time, since it will default to using
`delete` and will correctly release what's being new()-ed.

This pattern effectively replace the vlc_obj_malloc/free() primitive in
C++ code.