text_renderer: sapi: move to dedicated COM thread
Move the Speech synthesis API calls into a dedicated MTA COM thread so
that it's always guaranteed that those objects live in such thread.
We use a producer-consumer fifo of one slot to prevent storing the text
regions inside the COM thread, and in particular prevent duplicating
them, through we currently keep duplicating the last text rendered for
simplicity.

The setup feels weird, given that we give a single instruction to a
dedicated thread and then wait for that thread to process it. But
everything boils down to thread COM model being global to a thread and
not local to a scope, which means that we must control in which thread
the allocated COM objects will be used and released (ie. the same as
the one creating them).