avcodec/chroma: prioritize RGB mapping with the same mask
If the mask corresponds exactly, we have an exact match. If the RGB source
has a mask but none of the RGB mapping uses the same mask, we map to
a close RGB format with a different mask.

It's better than using the first VLC_CODEC_RGBxx in the list if there's
one corresponding to the exact match.

Another option would be to move the VLC_RGB_ES() calls at the end of the
list.