access/dshow: use VLC_CODEC_BGRX for MEDIASUBTYPE_RGB32
This is the proper type corresponding to MEDIASUBTYPE_RGB32 [1].
We don't need to handle VLC_CODEC_RGB32 anymore.

The use of VLC_CODEC_RGBA for MEDIASUBTYPE_ARGB32 is suspicious considering [1],
it should probably be VLC_CODEC_BGRA.

[1] https://learn.microsoft.com/en-us/windows/win32/directshow/uncompressed-rgb-video-subtypes
"If the subtype is MEDIASUBTYPE_ARGB32, byte 3 contains a value for the
alpha channel. If the subtype is MEDIASUBTYPE_RGB32, byte 3 should be ignored."