xwd: replace VLC_CODEC_RGB32 with VLC_CODEC_XRGB
This is the equivalent of VLC_CODEC_ARGB used below, without the alpha.

The TODO on RGB mask in ce0db74040b472094ac77762ade60bd466903233 means
they not have been known when it was written. The new code is not the
code that could have been assume (BGRX on Little Endian).