test/opengl: sub_renderer: add test for orientation
The test draws a 2x2 subpicture region on top of a 4x4 framebuffer. Each
pixel is different on the subpicture, which allows to check the
orientation and position of the subpicture according to the surface
orientation.

libplacebo is linked into the tests for now, since it seems the tests
need it transitively, but it should not be needed in the long run. See
the link error below:

    /usr/bin/ld: ../modules/video_output/opengl/test_modules_video_output_opengl_sub_renderer-sub_renderer.o:(.data.rel.ro+0x8): undefined reference to `pl_filter_spline16'
    /usr/bin/ld: ../modules/video_output/opengl/test_modules_video_output_opengl_sub_renderer-sub_renderer.o:(.data.rel.ro+0x10): undefined reference to `pl_filter_spline36'
    /usr/bin/ld: ../modules/video_output/opengl/test_modules_video_output_opengl_sub_renderer-sub_renderer.o:(.data.rel.ro+0x18): undefined reference to `pl_filter_spline64'
    /usr/bin/ld: ../modules/video_output/opengl/test_modules_video_output_opengl_sub_renderer-sub_renderer.o:(.data.rel.ro+0x20): undefined reference to `pl_filter_mitchell'
    /usr/bin/ld: ../modules/video_output/opengl/test_modules_video_output_opengl_sub_renderer-sub_renderer.o:(.data.rel.ro+0x28): undefined reference to `pl_filter_bicubic'
    ...