contrib: upnp: update to 1.14.18
To include the following commit that was causing high CPU usage on
vlc-ios:

miniserver: fix busy loop on socket error

In case of a socket error, the socket was not removed from the select
pool, causing future select() calls to return immediately, and
readFromSSDPSocket() to be called just after on the failing socket. This
was causing a high CPU load.