vlc_player: add a new on_stopping_current_media event
Add a new event notifying when the current media is stopping.

The media can be stopping without the player being stopping, ie. when
the media is finished and the player prepares the next one, but also
when the player changes media while having the player in the
VLC_PLAYER_STATE_PLAYING state.

The event will also be fired when the player is simply stopping, in
which case we expect that VLC_PLAYER_STATE_STOPPING is signaled before
the input is actually stop, which is coherent with the current
behaviour.