lib: media_player: split function
Splitting the function allows return statement, which removes the need
for break, and simplify the code. It also allows to use a designated
initializer for the event structure.

The code also move the vlc_assert_unreachable in the new function.

By moving the vlc_assert_unreachable after the switch(), we can get
warnings when not every cases are handled since there is no default
case, and we can keep asserting when it fails to reach any return
statement before.