demux: subtitle: refactor timing parsing
Refactor the parsing of timing values so that it doesn't allocate
memory. Indeed, we just need to parse the input string to locate where
the delimiter is, and call the parsing of timing values on each side of
the arrow.

We also use a length validation defensively, despite the current
impossibility to parse characters like `-->` in the sscanf given the
current format, to ensure that future addition here would not try to
parse additional characters and to ensure that future usage of the
function on strings that were not splitted by such character before
(eg. 0:0:0, 0 limited on 6 characters) don't impact the parsing.

(adapted from commit c76b8a42d216382b12e2de5d9cfd4d34a04a7431)