demux: subtitle: improve parsing of sized string
As mentioned in previous commit, it's possible for the value parsing
function to be called on a shorter string of a longer valid format,
which should not be parsed in that case.

The problem didn't arise previously because the function could only be
called on null-terminated strings and those strings were only built from
the split of ` --> ` into two strings.

(cherry picked from commit c74180ad548326e2c7537c0e0d0df19aecdde3af)