libvlc_events: add libvlc_MediaPlayerMediaStopping
The event forwards the vlc_player_t event `on_stopping_current_media`
which signals that the media actually playing inside the player is
stopping.

The event is essential to the `libvlc_media_new_callbacks` API since
there's no way to stop the \ref libvlc_media_read_cb callback without
triggering an EOF, and the EOF should be signalled only as soon as the
media is stopping.