lib: media player: signal stopping media event
Signal that the media is stopping to the libvlc client. This is done
after notifying the media on the VLC side that we're going to stop it,
so that application can rely on the fact that the objects using
application resources are already stopping, so that we can trigger
actions like EOS, closing those resources, etc.