input: don't signal EOF when input is stopped
When stopping an input, the player is signalled with vlc_player_Stop(),
which set the input for destruction in the player destructor thread.

When we called input_Stop(), the input thread will continue running, but
the demux will stop and notify the EOF state. This is reported by the
modified code to the es_out which considers it as a drain request, and
it's draining the decoders and the outputs despite the request to stop.

When calling input_Stop(), we are guaranteed that input_Stopped() will
be true afterwards, thus removing the signal removes this drain, and
removing the EOF signalling makes sense if the media was not finished
and was stopped.