input: es_out: early return
The if else block is followed by a single return, but the else block is
much more complex than the if block. To simplify, we can add the early
return to the first branch as an early success, ie. there is no other
handling necessary for PCR if buffering is still enabled.