test: add test for input clock
Add a unit test checking the behaviour of the input clock. For now, the
test mainly setup the infrastructure for testing the input clock and
checks some basic fundamentals.

Note that the test doesn't pass without the previous commit removing the
CR_PCR_GAP hack for DVD, so the test actually still check something that
was failing before.