bitmapinfoheader: fix double hton32 swap on masks
We do a first hton32 on the mask and then SetDWBE does it again.
What we really want is to store the mask as Little-Endian. This is also coherent
with the GetDWLE() used for reading the masks in ParseBitmapInfoHeader().

The corresponds to the 3 colors in extra data after the VLC_BITMAPINFOHEADER
for BI_BITFIELD [1]:

> If biCompression equals BI_BITFIELDS, the bitmap uses three DWORD color
> masks (red, green, and blue, respectively), which specify the byte layout
> of the pixels. The 1 bits in each mask indicate the bits for that color
> within the pixel.

The "bitmap" being the data after the "bitmap header".

[1] https://learn.microsoft.com/en-us/windows/win32/api/wingdi/ns-wingdi-bitmapinfoheader